\name{EPPS-package}
\alias{EPPS-package}
\alias{EPPS}
\docType{package}
\title{
\packageTitle{EPPS}
}
\description{
\packageDescription{EPPS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EPPS}
\packageIndices{EPPS}
~~ An overview of how to use the ~~
~~ package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{EPPS}

Maintainer: \packageMaintainer{EPPS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard ~~
~~ keywords, one per line, from ~~
~~ file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
